requirements.txt
Django==5.2.4
djangorestframework==3.16.0
requests==2.32.4
django-cors-headers==4.9.0
gunicorn==25.0.1
uuid6==2025.0.1
python-dotenv==1.1.1